Removes ChatBottomSheet from input field
Removes the ChatBottomSheet from the chat input field to relocate it to its own dedicated file.

This change improves code organization and maintainability by isolating the bottom sheet component.
